final
Okay so, adds upload button next to the delete button in the upcoming events, brings up a submit modal, click it. First time it will post the event embed (like the bot used to do) and any other time you click it, it will edit it instead so it doesnt send a new one instead.